+++ src/main/java/com/icthh/xm/gate/domain/idp/IdpConfigContainer.java
@@ -1,7 +1,7 @@
 package com.icthh.xm.gate.domain.idp;
 
-import com.icthh.xm.commons.domain.idp.model.IdpPrivateConfig.IdpConfigContainer.IdpPrivateClientConfig;
-import com.icthh.xm.commons.domain.idp.model.IdpPublicConfig.IdpConfigContainer.IdpPublicClientConfig;
+import com.icthh.xm.commons.domain.idp.IdpPrivateConfig.IdpConfigContainer.IdpPrivateClientConfig;
+import com.icthh.xm.commons.domain.idp.IdpPublicConfig.IdpConfigContainer.IdpPublicClientConfig;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
